fix: send current players at room message change: reorganizing packages from presentation messages
