+++ server/server-engine/src/main/java/com/thegame/server/engine/messages/output/AreaMessageBean.java
@@ -1,10 +1,13 @@
 package com.thegame.server.engine.messages.output;
 
+import com.owlike.genson.annotation.JsonIgnore;
 import com.thegame.server.engine.messages.IsMessageBean;
+import java.util.List;
 import java.util.Map;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
 import lombok.Singular;
@@ -26,4 +29,6 @@ public class AreaMessageBean implements IsMessageBean{
 	
 	@Singular
 	private Map<String,String> exits;
+	@Getter(onMethod_={@JsonIgnore})		
+	private List<String> players;
 }
