+++ server/server-presentation/src/main/java/com/thegame/server/presentation/endpoints/TheGameEndpoint.java
@@ -8,9 +8,9 @@ import com.thegame.server.engine.messages.input.RegisterPlayerMessageBean;
 import com.thegame.server.engine.messages.input.UnregisterPlayerMessageBean;
 import com.thegame.server.presentation.exceptions.PresentationException;
 import com.thegame.server.presentation.exceptions.PresentationExceptionType;
-import com.thegame.server.presentation.messages.ErrorMessage;
+import com.thegame.server.presentation.messages.output.ErrorMessage;
 import com.thegame.server.presentation.messages.IsMessage;
-import com.thegame.server.presentation.messages.support.MessageFactory;
+import com.thegame.server.presentation.support.MessageFactory;
 import java.time.LocalDateTime;
 import java.util.Optional;
 import java.util.stream.Stream;
