+++ server/server-presentation/src/main/java/com/thegame/server/presentation/messages/input/MoveMessage.java
@@ -1,5 +1,6 @@
-package com.thegame.server.presentation.messages;
+package com.thegame.server.presentation.messages.input;
 
+import com.thegame.server.presentation.messages.IsMessage;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
