+++ server/server-presentation/src/test/java/com/thegame/server/presentation/messages/ChatMessageTest.java
@@ -1,5 +1,6 @@
 package com.thegame.server.presentation.messages;
 
+import com.thegame.server.presentation.messages.input.ChatMessage;
 import javax.websocket.DecodeException;
 import javax.websocket.EncodeException;
 import org.junit.Test;
